Don't extract the default column value twice
The column default setting, not its value, should be used when creating
a new PostgreSQLColumn because PostgreSQLColumn#initialize calls
extract_value_from_default. When called on an already extracted default
value, the result is a nil value being used instead of the actual default
value defined in the DDL, thus leading to INSERT errors when the column
has a not-NULL constraint and no value is specifed when the record is
created.